+++ androidx/src/androidTest/java/me/panpf/androidx/test/hardware/HardwareTest.java
@@ -57,7 +57,7 @@ public class HardwareTest {
     @Test
     public final void testDeviceId() {
         Context context = InstrumentationRegistry.getContext();
-        if (Permissionx.isGrantPermission(context, Manifest.permission.READ_PHONE_STATE)) {
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
             String deviceId = Hardwarex.getDeviceId(context);
             Assert.assertTrue(deviceId.length() > 0 && !Premisex.areEqual(deviceId, "unknown") && !Premisex.areEqual(deviceId, "PermissionDenied"));
         } else {
@@ -73,7 +73,7 @@ public class HardwareTest {
     @Test
     public final void testSubscriberId() {
         Context context = InstrumentationRegistry.getContext();
-        if (Permissionx.isGrantPermission(context, Manifest.permission.READ_PHONE_STATE)) {
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
             String subscriberId = Hardwarex.getSubscriberId(context);
             Assert.assertTrue(subscriberId.length() > 0 && !Premisex.areEqual(subscriberId, "unknown") && !Premisex.areEqual(subscriberId, "PermissionDenied"));
         } else {
@@ -85,7 +85,7 @@ public class HardwareTest {
     @Test
     public final void testSimSerialNumber() {
         Context context = InstrumentationRegistry.getContext();
-        if (Permissionx.isGrantPermission(context, Manifest.permission.READ_PHONE_STATE)) {
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
             String simSerialNumber = Hardwarex.getSimSerialNumber(context);
             Assert.assertTrue(simSerialNumber.length() > 0 && !Premisex.areEqual(simSerialNumber, "unknown") && !Premisex.areEqual(simSerialNumber, "PermissionDenied"));
         } else {
@@ -97,7 +97,7 @@ public class HardwareTest {
     public final void testSerial() {
         if (Androidx.isAtLeastP()) {
             Context context = InstrumentationRegistry.getContext();
-            if (Permissionx.isGrantPermission(context, Manifest.permission.READ_PHONE_STATE)) {
+            if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
                 String serial = Hardwarex.getSerial();
                 Assert.assertTrue(serial.length() > 0 && !Premisex.areEqual(serial, "unknown") && !Premisex.areEqual(serial, "PermissionDenied"));
             } else {
@@ -111,7 +111,7 @@ public class HardwareTest {
     @Test
     public final void testMacAddress() {
         Context context = InstrumentationRegistry.getContext();
-        if (Permissionx.isGrantPermission(context, Manifest.permission.ACCESS_WIFI_STATE)) {
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.ACCESS_WIFI_STATE)) {
             String macAddress = Hardwarex.getMacAddress(context);
             Assert.assertTrue(macAddress.length() > 0 && !Premisex.areEqual(macAddress, "unknown") && !Premisex.areEqual(macAddress, "PermissionDenied"));
         } else {