+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fragment add instance method
 * Json 增加 format 方法
 * Window 增加 hasNavigationBar getStatusBarHeight getNavigationBarWidth 方法
+* Permission 增加 filterDeniedPermissions
 
 ## v0.1
 * 初次发布